Consolidate settings
